[#129] Reorganize modules according to versioning
Problem: with the current modules structure attempt to add V2 contract
would turn the codebase into mess.

Solution: separate things in three groups:
* Everything persistent or common between all versions of the contract
  goes to `Common` modules.
* Initial contract implementation goes to `V0` module.
* Everything specific to V1, what we may want to change in new versions,
  goes to `V1` modules.
