+++ src/Lorentz/Contracts/TZBTC.hs
@@ -37,10 +37,11 @@ module Lorentz.Contracts.TZBTC
 import Lorentz
 
 import Lorentz.Contracts.Metadata
-import Lorentz.Contracts.TZBTC.FlatParameter
+import Lorentz.Contracts.TZBTC.Common.FlatParameter
 import Lorentz.Contracts.TZBTC.Preprocess
-import Lorentz.Contracts.TZBTC.Types as Types
+import Lorentz.Contracts.TZBTC.Types
 import Lorentz.Contracts.TZBTC.V0
+import Lorentz.Contracts.TZBTC.V1
 
 -- Implementation
 ----------------------------------------------------------------------------
