+++ src/Lorentz/Contracts/TZBTC/Common/FlatParameter.hs
@@ -5,7 +5,7 @@
 {-# LANGUAGE RebindableSyntax #-}
 {-# OPTIONS_GHC -fno-warn-unused-do-bind #-}
 
-module Lorentz.Contracts.TZBTC.FlatParameter
+module Lorentz.Contracts.TZBTC.Common.FlatParameter
   ( FlatParameter(..)
   , fromFlatParameter
   ) where
@@ -17,11 +17,10 @@ import Lorentz.Contracts.Metadata
 import Lorentz.Contracts.Upgradeable.Common hiding (Parameter(..), Storage, mkEmptyStorage)
 import Util.Named
 
-import Lorentz.Contracts.TZBTC.Types
+import Lorentz.Contracts.TZBTC.Common.Types
   (AcceptOwnershipParams, ApproveParams, BurnParams, GetAllowanceParams, GetBalanceParams,
-  MintParams, OperatorParams, SetRedeemAddressParams,
-  TransferOwnershipParams, TransferParams)
-import qualified Lorentz.Contracts.TZBTC.Types as TZBTC
+  MintParams, OperatorParams, SetRedeemAddressParams, TransferOwnershipParams, TransferParams)
+import qualified Lorentz.Contracts.TZBTC.Common.Types as TZBTC
 
 -- | This is a type that is meant to hide details of how the actual parameter
 -- is structured, so that consumers of the contract are spared of any changes
